Create SMAPI.bat
Add new SMAPI wrapper. This will create StardewModdingAPI.deps.json if it doesn't exist, avoiding a Wabbajack build error, where the contents of this file exactly match a game data file, that is not identical between GoG and Steam releases.

It will then launch SMAPI, and ensure that only TBV mods are loaded.